+++ lib/css.mli
@@ -2210,7 +2210,11 @@ type page_size = Properties.page_size =
   | Named of page_size_name
   | Named_oriented of page_size_name * page_size_orientation
   | Oriented of page_size_orientation
+  | Initial
   | Inherit
+  | Unset
+  | Revert
+  | Revert_layer
   | Var of page_size var
 
 (** CSS columns values for multi-column layout. *)